Projects: switched parent to template
Each project is created from a template. The default one is the PRJ_LINT_SLUG.
I changed from parent to template because after creation there are no more relations
or parents' updates propagation...

Parent usually means that there is a strict relation with children,
and in OOP parent means that a child can be "partial": things not defined
in child are retrieved dynamically from parent.

This is not the case of a JAGOM project template, in which the template
acts only as a "starting point" for the new project.

Without any doubt can be useful to notify templates' updates to project instances.